requirements.txt
# requirements.txt
cryptography>=42.0
python-nmap>=0.7
scapy>=2.5
smbprotocol>=1.10
cryptography